Use double precision for minimizer settings

+++ interface/BayesianToyMC.h
@@ -29,7 +29,7 @@ private:
   /// number of toy mc computations to run
   static unsigned int tries_;
   /// Safety factor for hint (integrate up to this number of times the hinted limit)
-  static float hintSafetyFactor_;
+  static double hintSafetyFactor_;
 
   static std::vector<std::string> twoPoints_;
   std::pair<double,double> priorPredictiveDistribution(RooStats::ModelConfig *mc, RooAbsData &data, const RooArgSet *point=0, double *offset=0);
